Refactor agent manager log collection into workload log facade
